- Move server_method registration into portable-class machinery; support decorated classmethods; enforce decorator order; clean up type info - Add model classes and buffered row changes in Data Tables (draft rows, buffered mode); client/server capability flags (client_writable/creatable/updatable/deletable) [unsure—tables feature but runtime-related] - Fix PhoneNumberValidator handling of **properties - Support debugging in M3 app event handlers (Skulpt update adds logging module, itertools.batched/pairwise) - Pluggable UI modal: allow overriding default modal implementation (alerts/confirm) for advanced use cases - Refresh data bindings when reusing a layout (fixes titles/props not updating on navigation) - Fix DataGrid auto_header toggling when nested grids are present - Raise clearer errors on client for server-only anvil.server APIs (e.g., list_background_tasks) - Fix custom component instances sharing default mutable values (e.g., text[] default list) - Merge Skulpt upstream to remove string interning memory leak; note that `is` vs literals may no longer match (use `==`) - Fix signup_with_facebook - Make accelerated row.fetch(q.fetch_only(...)) public to refresh specific columns [unsure—tables feature but used via runtime] - Disallow model classes overriding __new__/__init__ (instantiation path is special) [unsure—tables/model but runtime-adjacent] - Relax autocreate-columns transaction so buffered .save() doesn’t fail when adding new columns [unsure—tables/runtime boundary] - Uplink: fix exceptions when app has breakpoints during uplink server calls [unsure—uplink vs runtime]
Based-on: anvil c58ee5accc8fa7875dbf42febfbd8ed573fec6fd

+++ database/migrator-core/resources/migrations/data-tables/schema.sql
@@ -1,6 +1,9 @@
 -- This is the schema that will set up a new Postgres data-tables database for an Anvil server
 -- This is required for data tables dbs, dedicated dbs, and combined central dbs. 
 -- Central db's being run along side a data tables db, do not include this schema.
+-- This file is not used by the migrator, but is guaranteed to be correct
+-- by the `platform/database/migrator/tests/check_migrations.py` script.
+
 -- DO NOT run this to upgrade an existing server.
 
 CREATE TABLE app_storage_tables (id integer primary key, 
@@ -14,7 +17,7 @@ CREATE TABLE app_storage_access (table_id integer references app_storage_tables(
 
 ALTER TABLE app_storage_access ADD CONSTRAINT app_storage_access_unique_table_id UNIQUE(table_id);
 
-CREATE TABLE app_storage_data (id serial, 
+CREATE TABLE app_storage_data (id bigserial, 
                                table_id integer references app_storage_tables(id), 
                                data jsonb NOT NULL DEFAULT '{}');
 CREATE INDEX app_storage_data_idx ON app_storage_data (table_id,id);
@@ -23,7 +26,7 @@ CREATE INDEX app_storage_data_json_idx ON app_storage_data using gin (data jsonb
 CREATE TABLE app_storage_media (object_id integer PRIMARY KEY NOT NULL,
                                 content_type text,
                                 name text,
-                                row_id integer,
+                                row_id bigint,
                                 table_id integer,
                                 column_id text,
                                 data bytea);
